requirements.txt
matplotlib==3.7.2
pandas==2.0.3
streamlit==1.27.2
tabulate==0.9.0
reportlab==3.6.2
Pillow==9.0.0
PyMuPDF==1.18.16
